# ghayath/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: "GHAYATH PERSONAL AI - Core API"
  version: "1.1.0"
  description: |
    REST/JSON API contract for **GHAYATH PERSONAL AI** (v1.1) — a Modular Agentic
    Architecture where an autonomous agent monitors and operates the owner's projects,
    inboxes, and communications under strict permission + approval control.

    ## Global rules

    **Envelope.** Every response (success or error) uses exactly the standard envelope:

    ```json
    { "success": true,  "data": { }, "request_id": "req_01J...", "timestamp": "2026-09-11T20:00:00Z" }
    { "success": false, "error": { "code": "PERMISSION_DENIED", "message": "...", "details": { } },
      "request_id": "req_01J...", "timestamp": "2026-09-11T20:00:00Z" }
    ```

    **Time.** All timestamps are ISO-8601 UTC (`...Z`) internally. Clients convert to
    the display timezone. Dates use `YYYY-MM-DD`.

    **IDs.** String identifiers with domain prefixes (ULID-based, human-greppable,
    log-safe): `usr_`, `prj_`, `ts_`, `mem_`, `cmd_`, `plan_`, `exec_`, `appr_`,
    `auto_`, `notf_`, `inc_`, `conv_`, `msg_` (email), `wmsg_` (WhatsApp),
    `draft_`, `repo_`, `iss_`, `evt_`, `req_`, `chk_`.

    **Authentication.** `Authorization: Bearer <JWT>` on every operation except
    `POST /auth/login`, `GET /health`, and `POST /webhooks/whatsapp`
    (signed via `X-Webhook-Signature`, HMAC-SHA256 of the raw body).
    Roles (RBAC): `OWNER` (full control), `AGENT` (service principal of the internal
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      tags:
      - Notifications
      operationId: getNotifications
      summary: Notification archive
      description: 'OWNER, VIEWER. Newest first. `limit` (default 50, max 200) is the

        documented exception from the no-pagination v1 list style — the archive

        grows unbounded.

        '
      parameters:
      - name: severity
        in: query
        required: false
        schema:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - QUEUED
          - SENT
          - FAILED
          - DISMISSED
      - name: project_id
        in: query
        required: false
        schema:
          type: string
      - name: since
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeNotificationList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Notifications]
      operationId: createNotification
      summary: Emit a notification
      description: |
        Severity-based routing when `channel` is omitted:
        `CRITICAL → PUSH + WHATSAPP`, `HIGH → PUSH`, `MEDIUM → DAILY_BRIEF`,
        `LOW → DASHBOARD`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
            example:
              severity: CRITICAL
              channel: PUSH
              title: Production Down
              message: SANAA production endpoint is unavailable
      responses:
        '201':
          description: Notification accepted and routed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/IntegrationOffline'

  # ─────────────────────────── AUDIT ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by `POST /auth/login` (1h) — refresh via `POST /auth/refresh`.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: HMAC-SHA256 hex digest of the raw request body, keyed with the provider's shared webhook secret.
...
  headers:
...
    XRateLimitReset:
      description: Unix epoch seconds at which the rate-limit window resets.
      schema:
        type: integer
        minimum: 0
  responses:
    BadRequest:
      description: Malformed request or failed validation (`VALIDATION_ERROR`; field problems in `error.details.fields`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing/invalid/expired credentials (`AUTHENTICATION_FAILED` or `INVALID_TOKEN`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not permitted (`PERMISSION_DENIED`; required permission in `error.details.required`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded (`RATE_LIMITED`).
      headers:
        X-RateLimit-Reset:
          $ref: '#/components/headers/XRateLimitReset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_LIMITED
              message: "Rate limit exceeded, retry after the reset"
              details: {}
            request_id: req_01J9ZK7Q3M4N5P6R7S8T9V
            timestamp: "2026-09-11T20:00:00Z"
    IntegrationOffline:
      description: External provider not connected/unreachable (`INTEGRATION_OFFLINE`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Base of every successful response.
      required: [success, request_id, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        request_id:
          type: string
          description: Unique correlation id assigned by the gateway.
          example: req_01J9ZK7Q3M4N5P6R7S8T9V
        timestamp:
          type: string
          format: date-time
          description: Server response time, ISO-8601 UTC.
          example: "2026-09-11T20:00:00Z"
    ErrorCode:
      type: string
      description: Fixed error-code set. Do not introduce new codes without a spec revision.
      enum:
        - AUTHENTICATION_FAILED
        - INVALID_TOKEN
        - PERMISSION_DENIED
        - APPROVAL_REQUIRED
        - VALIDATION_ERROR
        - RESOURCE_NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTEGRATION_OFFLINE
        - TOOL_UNAVAILABLE
        - EXECUTION_FAILED
        - VERIFICATION_FAILED
        - TIMEOUT
        - DEPENDENCY_FAILURE
        - SECRET_ACCESS_DENIED
    ErrorResponse:
      type: object
      description: Base of every error response.
      required: [success, error, request_id, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: Human-readable message (may be localized).
            details:
              type: object
              additionalProperties: true
              description: Structured context — field errors, approval_id, required permission, policy name, ...
        request_id:
          type: string
          example: req_01J9ZK7Q3M4N5P6R7S8T9V
        timestamp:
          type: string
          format: date-time
          example: "2026-09-11T20:00:00Z"

    # ─────────── Shared enums ───────────
...
    NotificationSeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    NotificationChannel:
      type: string
      enum: [PUSH, WHATSAPP, EMAIL, DASHBOARD, DAILY_BRIEF]
    CreateNotificationRequest:
      type: object
      required: [severity, title, message]
      properties:
        severity:
          $ref: '#/components/schemas/NotificationSeverity'
        channel:
          type: string
          enum: [PUSH, WHATSAPP, EMAIL, DASHBOARD]
          description: "Preferred channel. Omit to let severity-based routing decide: CRITICAL→PUSH+WHATSAPP, HIGH→PUSH, MEDIUM→DAILY_BRIEF, LOW→DASHBOARD."
        title:
          type: string
          maxLength: 200
          example: Production Down
        message:
          type: string
          maxLength: 4000
          example: SANAA production endpoint is unavailable
        project_id:
          anyOf:
            - type: string
              pattern: '^prj_[a-z0-9]+$'
            - type: 'null'
        data:
          type: object
          additionalProperties: true
    CreateNotificationData:
      type: object
      required: [notification_id, routed_channels, status]
      properties:
        notification_id:
          type: string
          pattern: '^notf_[a-z0-9]+$'
        routed_channels:
          type: array
          items:
            $ref: '#/components/schemas/NotificationChannel'
        status:
          type: string
          enum: [QUEUED, SENT, FAILED, DISMISSED]
    NotificationResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/CreateNotificationData'

    # ─────────── Audit ───────────
...
    NotificationView:
      type: object
      required:
      - id
      - severity
      - status
      - title
      - message
      - created_at
      properties:
        id:
          type: string
        severity:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
        channel_requested:
          type: string
          enum:
          - PUSH
          - WHATSAPP
          - EMAIL
          - DASHBOARD
        channels:
          type: array
          items:
            type: string
        title:
          type: string
        message:
          type: string
        project_id:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - QUEUED
          - SENT
          - FAILED
          - DISMISSED
        data:
          type: object
        created_at:
          type: string
          format: date-time
        sent_at:
          type:
          - string
          - 'null'
          format: date-time
    EnvelopeNotificationList:
      allOf:
      - $ref: '#/components/schemas/SuccessEnvelope'
      - type: object
        properties:
          data:
            type: array
            items:
              $ref: '#/components/schemas/NotificationView'